pyproject.toml
[tool.poetry]
name = "service-registry"
version = "0.1.0"
description = ""
authors = ["Olga Lipina"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.3"
redis = "^5.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
